backend/requirements.txt
botocore==1.40.48
cryptography==46.0.2
Flask==3.1.2
Flask_Cors==5.0.0
pandas==2.3.3
psycopg2==2.9.10
psycopg2_binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
bcrypt>=4.3.0
Requests==2.32.5
sib_api_v3_sdk==7.6.0
Werkzeug==3.1.3
gunicorn==22.0.0